requirements.txt
click==8.1.7
hydra-core==1.3.2
matplotlib==3.8.2
omegaconf==2.3.0
scikit_learn==1.3.2
torch==2.1.2
torchvision==0.16.2
